+++ vkconfig_core/test/test_executable.cpp
@@ -130,13 +130,17 @@ TEST(test_executable, GetDefaultExecutablePath) {
     std::string saved = qgetenv("VULKAN_SDK").toStdString();
     qunsetenv("VULKAN_SDK");
 
-    DefaultExecutable default_executable{"vkcube", "/vkcube", "--suppress_popups", "vkcube launcher options"};
-    const DefaultPath& default_paths = ::GetDefaultExecutablePath(default_executable.key);
 #if defined(_WIN32)
+    DefaultExecutable default_executable{"/vkcube.exe", "--suppress_popups", "vkcube launcher options"};
+    const DefaultPath& default_paths = ::GetDefaultExecutablePath(default_executable.key);
     EXPECT_STREQ(default_paths.executable_path.RelativePath().c_str(), Path(".\\vkcube.exe").RelativePath().c_str());
 #elif defined(__linux__)
+    DefaultExecutable default_executable{"/vkcube", "--suppress_popups", "vkcube launcher options"};
+    const DefaultPath& default_paths = ::GetDefaultExecutablePath(default_executable.key);
     EXPECT_STREQ(default_paths.executable_path.RelativePath().c_str(), Path(".\\vkcube").RelativePath().c_str());
 #elif defined(__APPLE__)
+    DefaultExecutable default_executable{"/vkcube.app", "--suppress_popups", "vkcube launcher options"};
+    const DefaultPath& default_paths = ::GetDefaultExecutablePath(default_executable.key);
     EXPECT_STREQ(default_paths.executable_path.RelativePath().c_str(),
                  Path("/Applications/vkcube.app/Contents/MacOS/vkcube").RelativePath().c_str());
 #endif
@@ -155,18 +159,22 @@ TEST(test_executable, DefaultExecutable) {
     std::string saved = qgetenv("VULKAN_SDK").toStdString();
     qunsetenv("VULKAN_SDK");
 
-    DefaultExecutable default_executable{"vkcube", "/vkcube", "--suppress_popups", "vkcube launcher options"};
-    Executable executable(default_executable);
-
-    EXPECT_EQ(1, executable.GetOptions().size());
 #if defined(_WIN32)
+    DefaultExecutable default_executable{"/vkcube.exe", "vkcube", "--suppress_popups", "vkcube launcher options"};
+    Executable executable(default_executable);
     EXPECT_STREQ(executable.path.RelativePath().c_str(), Path(".\\vkcube.exe").RelativePath().c_str());
 #elif defined(__linux__)
+    DefaultExecutable default_executable{"/vkcube", "vkcube", "--suppress_popups", "vkcube launcher options"};
+    Executable executable(default_executable);
     EXPECT_STREQ(executable.path.RelativePath().c_str(), Path(".\\vkcube").RelativePath().c_str());
 #elif defined(__APPLE__)
+    DefaultExecutable default_executable{"/vkcube.app", "vkcube", "--suppress_popups", "vkcube launcher options"};
+    Executable executable(default_executable);
     EXPECT_STREQ(executable.path.RelativePath().c_str(),
                  Path("/Applications/vkcube.app/Contents/MacOS/vkcube").RelativePath().c_str());
 #endif
+
+    EXPECT_EQ(1, executable.GetOptions().size());
     EXPECT_STREQ(executable.configuration.c_str(), "Validation");
     EXPECT_EQ(executable.enabled, true);
     EXPECT_STREQ(executable.GetActiveOptionsName().c_str(), "vkcube launcher options");